feat: #86 Divide the ToastViewModel - 뷰와 테스트뷰, 모델을 분리
