+++ FiveGuyes/FiveGuyes/Sources/Views/Components/ToastView.swift
@@ -8,66 +8,45 @@
 import SwiftUI
 
 struct ToastView: View {
-    var message: String
-    var body: some View {
-        VStack(spacing: 0) {
-            Spacer()
-            Text(message)
-                .font(.system(size: 16, weight: .semibold))
-                .padding(.horizontal, 20)
-                .padding(.vertical, 14)
-                .foregroundColor(.white)
-                .background(Color.gray)
-                .cornerRadius(16)
-        }
-    }
-}
-
-class ToastModel: ObservableObject {
-    @Published var message: String?
-    @Published var isVisible: Bool = false
-    private var timer: Timer?
+    @ObservedObject var viewModel: ToastViewModel
     
-    func showToast(message: String, duration: TimeInterval = 1.0) {
-        self.message = message
-        self.isVisible = true
-        
-        timer?.invalidate()
-        
-        timer = Timer.scheduledTimer(withTimeInterval: duration, repeats: false) { [weak self] _ in
-            DispatchQueue.main.async {
-                withAnimation {
-                    self?.isVisible = false
-                }
-                DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-                    self?.message = nil
-                }
-            }
-        }
-    }
-}
-
-    struct ToastTestView: View {
-        @StateObject private var toastViewModel = ToastModel()
-        
-        var body: some View {
-            ZStack {
-                VStack {
-                    Button("Show Toast") {
-                        toastViewModel.showToast(message: "오늘은 완독하는 마지막 날이에요!")
-                    }
-                }
-                
-                if let message = toastViewModel.message {
-                    ToastView(message: message)
-                        .opacity(toastViewModel.isVisible ? 1.0 : 0.0)
-                        .animation(.easeInOut(duration: 0.5), value: toastViewModel.isVisible)
+    var body: some View {
+        ZStack {
+            VStack {
+                Spacer()
+                if let message = viewModel.message, viewModel.isVisible {
+                    Text(message)
+                        .font(.system(size: 16, weight: .semibold))
+                        .padding(.horizontal, 20)
+                        .padding(.vertical, 14)
+                        .foregroundColor(.white)
+                        .background(Color.gray)
+                        .cornerRadius(16)
                         .transition(.opacity)
+                        .animation(.easeInOut(duration: 0.5), value: viewModel.isVisible)
                         .zIndex(1)
                 }
             }
         }
     }
+}
+
+struct ToastTestView: View {
+    @StateObject private var toastViewModel = ToastViewModel()
+    
+    var body: some View {
+        ZStack {
+            VStack(spacing: 20) {
+                Button {
+                    toastViewModel.showToast(message: "오늘은 완독하는 마지막날이에요!")
+                } label: {
+                    Text("Show Toast")
+                }
+            }
+            ToastView(viewModel: toastViewModel)
+        }
+    }
+}
 
 #Preview {
     ToastTestView()
